Modified total price to float Percentage test fixes

+++ src/main/java/com/promo/calculator/MultiPromoCalculator.java
@@ -9,16 +9,16 @@ import java.util.List;
 
 public class MultiPromoCalculator {
 
-    public int calculate(List<ComboPromotion> combos) {
-        int price = 0;
+    public float calculate(List<ComboPromotion> combos) {
+        float price = 0;
         for(ComboPromotion combo  : combos){
-            int reducedSKUPrice = 0;
+            float reducedSKUPrice = 0;
             MultiSKUPromotion promotion = combo.getPromotion() instanceof MultiSKUPromotion ? ((MultiSKUPromotion) combo.getPromotion()) : null;
             if(promotion.getDiscountType().equals(DiscountType.FIXED_PRICE)){
                 reducedSKUPrice = promotion.getValue();
             } else {
-                int itemTotalPrice = combo.getItems().stream().map(Item::getPrice).reduce(Integer::sum).orElse(0);
-                reducedSKUPrice = itemTotalPrice * (promotion.getValue() / 100);
+                float itemTotalPrice = combo.getItems().stream().map(Item::getPrice).reduce(Integer::sum).orElse(0);
+                reducedSKUPrice = itemTotalPrice * (Float.valueOf(promotion.getValue()) / 100);
             }
             combo.setTotalPrice(reducedSKUPrice);
             price = price + reducedSKUPrice;
